requirements.txt
llama-cpp-python
rich